mudei uns if/else por try/catch

+++ src/jogadores/JogadorSortudo.java
@@ -22,8 +22,8 @@ public class JogadorSortudo extends Jogador{
 			int corrigido = (soma(dado) + 6);
 			if(corrigido % 2 == 0) {
 				try {
-				dado[0].setFace(corrigido/2);
-				dado[1].setFace(corrigido/2);
+				dado[0].setFace((int)(corrigido/2));
+				dado[1].setFace((int)(corrigido/2));
 				}catch(InvalidFaceException ive) {
 					ive.printStackTrace();
 				}
